requirements.txt
attrs==21.4.0
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
dnspython==2.2.1
email-validator==1.2.1
filelock==3.7.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask==2.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
lxml==4.9.0
MarkupSafe==2.1.1
parsel==1.6.0
pip==22.1.2
Protego==0.2.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyotp==2.6.0
python-dotenv==0.20.0
queuelib==1.6.2
requests-file==1.5.1
requests==2.27.1
Scrapy==2.6.1
service-identity==21.1.0
setuptools==62.3.2
six==1.16.0
SQLAlchemy==1.4.37
tldextract==3.3.0
Twisted==22.4.0
typing-extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==2.1.2
wheel==0.37.1
WTForms==3.0.1
zipp==3.8.0
zope.interface==5.4.0
